+++ src/analyze/+mag/+process/Cast.m
@@ -1,12 +1,6 @@
 classdef Cast < mag.process.Step
 % CAST Apply data type cast.
 
-    properties (Dependent)
-        Name
-        Description
-        DetailedDescription
-    end
-
     properties
         % DATATYPE Data type to convert to.
         DataType (1, 1) string
@@ -23,18 +17,6 @@ classdef Cast < mag.process.Step
             end
 
             this.assignProperties(options);
-        end
-
-        function value = get.Name(~)
-            value = "Cast to Data Type";
-        end
-
-        function value = get.Description(this)
-            value = "Cast " + join(compose("""%s""", this.Variables), ", ") + " to """ + this.DataType + """.";
-        end
-
-        function value = get.DetailedDescription(this)
-            value = this.Description;
         end
 
         function data = apply(this, data, ~)