+++ src/analyze/+mag/+process/Range.m
@@ -1,12 +1,6 @@
 classdef Range < mag.process.Step
 % RANGE Apply scale factor based on range value.
 
-    properties (Dependent)
-        Name
-        Description
-        DetailedDescription
-    end
-
     properties (Constant)
         % SCALEFACTORS Scale factor for each supported range.
         ScaleFactors (1, 4) double = [2.13618, 0.072, 0.01854, 0.00453]